server/requirements.txt
Flask
Flask-Cors
werkzeug
supabase
python-dotenv
llama_parse
firecrawl-py
firecrawl